Split up Concurrent allocator into separate modules
